fix(wingler): solved dependency loop
So had to tweak some things here, in order to use hooks in inference, inference needed to be a hook, and as a hook it needed to be in the dependency array of the useEffect for wingler, but that created issues. To solve it, we create a ref and save the ref och the current version of 'handleInference' in the current. Now eslint is fine with it not being in the array.

- wakeWordCallback, inferenceCallback and switchToVTT are wrapped in useCallback and moved out from the useEffect
- intentHandlers is wrapped in useMemo
- handleAppCommands. playback wrapped in useCallback
- commented the useEffect in the bottom, this could be nice in the future, I need to test eventEmitters